Updates existing weight logic to work with floats from the Jimmy Johnson trade model

+++ src/com/ryanlaverick/Selection.java
@@ -1,5 +1,5 @@
 package com.ryanlaverick;
 
-public record Selection(int round, int pick, int weight) {
+public record Selection(int round, int pick, float weight) {
     //
 }
